+++ packages/eds-core-react/src/components/Typography/types.ts
@@ -5,16 +5,7 @@
 export type FontFamily = 'ui' | 'header'
 
 export type FontSize =
-  | 'xs'
-  | 'sm'
-  | 'md'
-  | 'lg'
-  | 'xl'
-  | '2xl'
-  | '3xl'
-  | '4xl'
-  | '5xl'
-  | '6xl'
+  'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl' | '4xl' | '5xl' | '6xl'
 
 export type LineHeight = 'default' | 'squished'